refactor: add `applyGetter` helper
