+++ .changes/accordion.md
@@ -1,4 +1,5 @@
---
-material-ui-iteractors: patch
---
+---
+material-ui-interactors: patch
+---
+
 Added interactor for [Accordion](https://material-ui.com/components/accordion/) component
